Added Web services section
I think that this section is important to understand the generality of the page generation mechanism. the View monad can be used as a general request-response mechanism.
Adding a simple set of combinators it is possible to create web services in an idiomatic way without the need of special constructions.